// flyteconsole/packages/zapp/console/src/components/Launch/LaunchForm/types.ts
export enum InputType {
  Binary = 'BINARY',
  Blob = 'BLOB',
  Boolean = 'BOOLEAN',
  Collection = 'COLLECTION',
  Datetime = 'DATETIME',
  Duration = 'DURATION',
  Error = 'ERROR',
  Enum = 'ENUM',
  Float = 'FLOAT',
  Integer = 'INTEGER',
  Map = 'MAP',
  None = 'NONE',
  Schema = 'SCHEMA',
  String = 'STRING',
  Struct = 'STRUCT',
  Union = 'Union',
  Unknown = 'UNKNOWN',
}

export interface InputTypeDefinition {
  literalType: LiteralType;
  type: InputType;
  subtype?: InputTypeDefinition;
  listOfSubTypes?: InputTypeDefinition[];
}

export interface BlobValue {
  dimensionality: BlobDimensionality | string;
  format?: string;
  uri: string;
}
// ...
export enum AuthRoleTypes {
  k8 = 'k8',
  IAM = 'IAM',
}

export interface AuthRoleMeta {
  helperText: string;
  inputLabel: string;
  label: string;
  value: RoleTypeValue;
  validate?: any;
  error?: any;
}

export type RoleTypeValue = keyof Admin.IAuthRole;